Package Lock Upgrade
I don't like updating unrelated files in my initial types commits, but this was already going to be done with having to install TypeScript and `@types/node` anyways, so might as well make it more discernible as it's own commit. I didn't actually upgrade any of the versions for the dependencies, just the lock file.