TS Build Step Types!
Note that this doesn't necessarily mean that things are written in TS, which is extra awesome in my opinion. To be able to write things in Vanilla JS, with JSDoc, but also still be able to generate full TS type definitions for the library itself, is just great. Not only that, but that you get type checking in your editor, of course, as well. The whole thing is just great how it all works interchangeably.

I changed the build target to ES5, because it seems like this was meant to target older Node versions, but I'm not completely sure if that's the case or not. Maybe things can move to ESM instead, once I fully move it to TS, but for the time being I thought the build output should be similar in composition to that of how the code currently appears to be built for.
https://www.learningtypescript.com/articles/why-increase-your-tsconfig-target