UUID Class Declaration
Fully moved over to a class type now! This will help with having things working more easily out of the box in TS as well.

I had to disable the original use of JSHint, as the output that TS started using for this build step wasn't liked by it. I think the way to go over time is to move things to ESLint/TSLint. I haven't delved into that yet, so I don't have any experience in working with stricter linting rules for formatting and things like that.